+++ dist_apps/cc_pull/gen_cuda.cuh
@@ -7,7 +7,7 @@
 #include "galois/runtime/cuda/DeviceSync.h"
 
 struct CUDA_Context : public CUDA_Context_Common {
-	struct CUDA_Context_Field<unsigned long long> comp_current;
+	struct CUDA_Context_Field<uint32_t> comp_current;
 };
 
 struct CUDA_Context *get_CUDA_context(int id) {
@@ -46,23 +46,23 @@ void bitset_comp_current_reset_cuda(struct CUDA_Context *ctx, size_t begin, size
   reset_bitset_field(&ctx->comp_current, begin, end);
 }
 
-unsigned long long get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID) {
-	unsigned long long *comp_current = ctx->comp_current.data.cpu_rd_ptr();
+uint32_t get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID) {
+	uint32_t *comp_current = ctx->comp_current.data.cpu_rd_ptr();
 	return comp_current[LID];
 }
 
-void set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
-	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+void set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
+	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	comp_current[LID] = v;
 }
 
-void add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
-	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+void add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
+	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	comp_current[LID] += v;
 }
 
-bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
-	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
+	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	if (comp_current[LID] > v) {
 		comp_current[LID] = v;
     return true;
@@ -70,43 +70,42 @@ bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned
   return false;
 }
 
-void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
-	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_current, from_id, v);
+void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
+	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_current, from_id, v);
 }
 
-void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
-	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_current, from_id, v);
+void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
+	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_current, from_id, v);
 }
 
-void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v, unsigned long long i) {
-	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_current, from_id, v, i);
+void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v, uint32_t i) {
+	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_current, from_id, v, i);
 }
 
-void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode, unsigned long long i) {
-	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
+void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode, uint32_t i) {
+	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
 }
 
-void batch_set_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<unsigned long long, sharedMirror, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<uint32_t, sharedMirror, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<unsigned long long, sharedMaster, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<uint32_t, sharedMaster, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<unsigned long long, sharedMaster, addOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<uint32_t, sharedMaster, addOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<unsigned long long, sharedMaster, minOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<uint32_t, sharedMaster, minOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
-
